1. In the release environment, delete `NSAssert` in the animation/animations API. 2. filter the `_UISystemBackgroundView` for iOS14

+++ TABAnimatedDemo/TABAnimatedDemo/ViewController/CollectionView/UserCode/Controller/NestCollectionViewController.m
@@ -62,7 +62,7 @@
         [array addObject:@"test.jpg"];
         [self.dataArray addObject:array];
         [self.collectionView registerClass:[NestCollectionViewCell class]
-                forCellWithReuseIdentifier:[NSString stringWithFormat:@"NestCollectionViewCell %ld",i]];
+                forCellWithReuseIdentifier:[NSString stringWithFormat:@"NestCollectionViewCell %ld",(long)i]];
     }
     [self.collectionView tab_endAnimationEaseOut];
 }
